requirements.txt
nextcord==3.1.1
yt-dlp==2024.8.6
PyNaCl==1.5.0
ffmpeg-python==0.2.0
python-dotenv==1.0.1
requests>=2.32.2
aiohttp>=3.12.0
setuptools>=65.0.0
wheel>=0.38.0
packaging>=21.0
pillow>=10.0.1
jsonschema==4.19.2
PyJWT==2.8.0
cryptography>=41.0.7
python-dateutil==2.8.2
PyYAML==6.0.1
